#10 - Dodanie DisplayAlert dla potwierdzenia usunięcia wydarznia

+++ EventPulse/Pages/EditEventPage.xaml.cs
@@ -1,4 +1,5 @@
 using EventPulse.Models;
+using EventPulse.Resources.Languages;
 using EventPulse.Services;
 
 namespace EventPulse.Pages;
@@ -20,9 +21,13 @@ public partial class EditEventPage : ContentPage
 
 	private async void OnDeleteEventClicked(object sender, EventArgs e)
 	{
-		await App.EventDb.DeleteEvent(int.Parse(_selectedEvent.Id));
+		bool answer = await DisplayAlert(AppResources.Delete_confirmation, AppResources.Delete_confirmation_msg, AppResources.Yes, AppResources.No);
 
-		await Navigation.PushAsync(new HomePage());
+		if (answer)
+		{
+			await App.EventDb.DeleteEvent(int.Parse(_selectedEvent.Id));
+			await Navigation.PushAsync(new HomePage());
+		}
 	}
 
 	private async void OnSaveChangesClicked(object sender, EventArgs e)
